% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output-env.R
\docType{data}
\name{outputs}
\alias{outputs}
\title{Environment for Outputs Tables}
\format{
An object of class \code{environment} of length 0.
}
\usage{
outputs
}
\description{
This environment is created to store the outputs generated by a Redatam program
execution.
The environment contains TABLES (Redatam outputs) generated from FREQ, AREALIST,
MATOP, TABLIST commands.
You can't modify objects contained in this environment.
}
\examples{
\dontrun{
dic<-redatam_open("c:/nmir/nmiresp.rxdb")
df<-redatam_query(dic,"tbl1<-freq person.sexo")
# Accessing the tbl1 object from environment
redatam::outputs$tbl1
# Accessing the output from Global Environment
print(tbl1)
}

}
\keyword{datasets}
