% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redatam-api.R
\name{redatam_create}
\alias{redatam_create}
\title{Create Redatam database}
\usage{
redatam_create(schema_name, config_name)
}
\arguments{
\item{schema_name}{Dictionary schema in json format}

\item{config_name}{Configuration in json format}
}
\value{
{No return value.}
}
\description{
Create a new Redatam database.
}
\examples{
\dontrun{
redatam_create("path/to/schema.json","path/to/config.json")
}
}
\author{
Jaime Salvador
}
