% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redatam-api.R
\name{redatam_destroy_}
\alias{redatam_destroy_}
\title{Destroy engine}
\usage{
redatam_destroy_()
}
\value{
{No return value.}
}
\description{
Destroy the Redatam engine.
This function is automatically called when the package is unloaded.
}
\examples{
\dontrun{
redatam_destroy_();
}
}
\author{
Jaime Salvador
}
