% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redatam-api.R
\name{redatam_entities}
\alias{redatam_entities}
\title{List entities}
\usage{
redatam_entities(dic)
}
\arguments{
\item{dic}{Dictionary ID (returned by 'redatam.open')}
}
\value{
Data frame that contains all the entities in the database.
}
\description{
List the entities in a databse.
}
\examples{
\dontrun{
dic<-redatam_open("path/to/rxdb")
redatam_entities(dic);
}
}
\author{
Jaime Salvador
}
