% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redatam-api.R
\name{redatam_info}
\alias{redatam_info}
\title{Get extended Redatam API information}
\usage{
redatam_info()
}
\value{
String with the Redatam extended information.
}
\description{
Returns extended information about datasources and datasets available.
}
\examples{
redatam_info()
}
\author{
Jaime Salvador
}
