% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redatam-api.R
\name{redatam_init_}
\alias{redatam_init_}
\title{Initialize engine}
\usage{
redatam_init_(pachageDir)
}
\arguments{
\item{pachageDir}{Relative path to Redatam runtime (dll, so or dynlib)}
}
\value{
{No return value.}
}
\description{
Initialize the Redatam engine.
This function is automatically called when the package is loaded.
}
\examples{
\dontrun{
redatam_init_("/engine");
}
}
\author{
Jaime Salvador
}
\keyword{internal}
