% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redatam-api.R
\name{redatam_internal_query}
\alias{redatam_internal_query}
\title{Execute a Redatam command from text}
\usage{
redatam_internal_query(dic, spc)
}
\arguments{
\item{dic}{Dictionary identifier}

\item{spc}{Program text in SPC format}
}
\value{
Raw dataset with al values: tot, na, mv, values.
}
\description{
Execute a Redatam command: TABLE or AREALIST.
}
\examples{
\dontrun{
dic<-redatam_open("path/to/rxdb")
df<-redatam_internal_query(dic,"freq person.sexo")
}
}
\author{
Jaime Salvador
}
