% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redatam-plugins.R
\name{redatam_load_plugin}
\alias{redatam_load_plugin}
\title{Load a custom plugin}
\usage{
redatam_load_plugin(filePath)
}
\arguments{
\item{filePath}{Full path to .dll (windows), .so (linux) or .dynlib (macOs)}
}
\value{
no return.
}
\description{
Load a custom plugin into internal registry.
This function is used to add a custom plugin that implements a data source
(to create database from custom format) or data set (to save outputs in
custom format).
}
\examples{
\dontrun{
redatam_load_plugin("path/to/custom/plugin")
}

}
\author{
Jaime Salvador
}
