% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redatam-api.R
\name{redatam_open}
\alias{redatam_open}
\title{Open dictionary}
\usage{
redatam_open(dictionary_name)
}
\arguments{
\item{dictionary_name}{Dictionary filename}
}
\value{
ID than can be used in functions to query data.
}
\description{
Open a REDATAM database.
This function returns an ID than can be used in functions to query data.
}
\examples{
\dontrun{
dic<-redatam_open("path/to/rxdb")
}
}
\author{
Jaime Salvador
}
