% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redatam-api.R
\name{redatam_save}
\alias{redatam_save}
\title{Save dictionary}
\usage{
redatam_save(dic, name = "")
}
\arguments{
\item{dic}{Dictionary identifier}

\item{name}{Dictionary filename}
}
\value{
{No return value.}
}
\description{
Save a REDATAM database.
This function  can be used to save a dictionary.
}
\examples{
\dontrun{
dic<-redatam_open("path/to/rxdb")
...
redatam_save(dic,"path/to/new/dictionary.rxdb")
}
}
\author{
Jaime Salvador
}
