% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redatam-api.R
\name{redatam_variables}
\alias{redatam_variables}
\title{List variables for entity}
\usage{
redatam_variables(dic, entity_name)
}
\arguments{
\item{dic}{Dictionary ID (returned by 'redatam.open')}

\item{entity_name}{Entity's name}
}
\value{
Data frame that contains all the variables from the \sQuote{entity_name}.
}
\description{
List the variables in an entity.
}
\examples{
\dontrun{
dic<-redatam_open("path/to/rxdb")
redatam_variables(dic, "person")
}
}
\author{
Jaime Salvador
}
