% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redatam-api.R
\name{redatam_version}
\alias{redatam_version}
\title{Get the Redatam API version}
\usage{
redatam_version()
}
\value{
String with the Redatam Engine version.
}
\description{
Returns the current version of the native Redatam API.
}
\examples{
redatam_version()
}
\author{
Jaime Salvador
}
