% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redatam-api.R
\name{redatam_internal_run}
\alias{redatam_internal_run}
\title{Execute a Redatam command from file}
\usage{
redatam_internal_run(dic, file_name)
}
\arguments{
\item{dic}{Dictionary identifier}

\item{file_name}{Program file name}
}
\value{
Raw dataset with all values: tot, na, mv, values.
}
\description{
Execute a Redatam command: TABLE or AREALIST.
}
\examples{
\dontrun{
dic<-redatam.open("rxdb")
df<-redatam.internal_run(dic,"/path/to/nmir_test.spc")
}
}
\author{
Jaime Salvador
}
